fix issue where order of conditional check could rarely cause a segfault
the c++ spec will guarantee this will no longer be an issue, as if the first condition is false, it does not check the second.

+++ src/StringTree/StringTree.cpp
@@ -143,7 +143,7 @@ namespace NailForge::StringTree {
                             hasResolvedSearchRange = true;
                         }
 
-                        if (__builtin_expect(accumulatedScorePassesThreshold && (currentDepth < (context.searchParams.maximumHitLength)), false)) {
+                        if (__builtin_expect(accumulatedScorePassesThreshold && (currentDepth < ((int32_t)context.searchParams.maximumHitLength)), false)) {
                             verifyDiagonalsPassingThreshold(context, thisSearchRange, resolvedSequencePositionList,
                                 thisSymbolModelPosition, maxScoreSeen, seedList, currentDepth);
